fixing parameter in ordering_process.h

+++ include/process/ordering_process.h
@@ -31,6 +31,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <cstddef>
 #include <cstdint>
 #include <stdbool.h>
+#include <string>
 using Edge_t = struct Edge_t;
 using II_t_rt = struct II_t_rt;
 #else
@@ -38,13 +39,15 @@ using II_t_rt = struct II_t_rt;
 #include <stdint.h>
 typedef struct Edge_t Edge_t;
 typedef struct II_t_rt II_t_rt;
+
 #endif
+
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-void pgr_process_ordering(const char*, int16_t, II_t_rt **, size_t *);
+void pgr_process_ordering(const char*, int, II_t_rt**, size_t*);
 
 #ifdef __cplusplus
 }
